+++ Data/__init__.py
@@ -1,3 +1,3 @@
-from .experiment_descriptor import ExperimentDescriptor, ExperimentType, Role
+from .experiment_descriptor import ExperimentDescriptor, ExperimentType
 from .metal_requirements import Metal, MetalUsage
 from .ns_info import NsInfo