+++ Data/experiment_descriptor.py
@@ -1,5 +1,5 @@
 from Helper import Serialize
-from typing import Dict, Tuple, List
+from typing import Dict, Tuple, List, Optional
 from enum import Enum, unique
 from datetime import datetime, timezone
 
@@ -10,19 +10,7 @@ class ExperimentType(Enum):
     Standard = 1
     Custom = 2
     MONROE = 3
-
-
-@unique
-class Role(Enum):
-    Master = 0
-    Slave = 1
-
-
-class DistributedSettings:
-    def __init__(self, data: Dict):
-        self.Role = Role[data['Role']]
-        self.SlavePlatform = None if self.Role == Role.Slave else data['SlavePlatform']
-        self.SlaveExperiment = None if self.Role == Role.Slave else data['SlaveExperiment']
+    Distributed = 4
 
 
 class ExperimentDescriptor:
@@ -33,21 +21,15 @@ class ExperimentDescriptor:
         if self.Valid:
             self.Type = ExperimentType[data['ExperimentType']]
             self.Identifier = f"{time}{self.Type.name}:{(','.join(self.TestCases))}-{(','.join(self.UEs))}"
-            self.Distributed = DistributedSettings(data) if data['Distributed'] else None
         else:
             self.Type = ExperimentType.Error
             self.Identifier = time
-            self.Distributed = None
-
-        # TODO: Atts to delete
-        self.Id = self.Platform = self.HasNsd = None
 
     @staticmethod
     def validate(data: Dict) -> Tuple[bool, List[str]]:
         keys = ['Version', 'ExperimentType', 'TestCases', 'UEs', 'Slice', 'NSs',
                 'ExclusiveExecution', 'Scenario', 'Automated', 'ReservationTime',
-                'Application', 'Parameters', 'Distributed', 'Role', 'SlavePlatform',
-                'SlaveExperiment', 'Extra']
+                'Application', 'Parameters', 'Remote', 'Extra']
         return Serialize.CheckKeys(data, *keys)
 
     @property
@@ -101,6 +83,14 @@ class ExperimentDescriptor:
     @property
     def Extra(self) -> Dict[str, object]:
         return self._data['Extra']
+
+    @property
+    def Remote(self) -> Optional[str]:
+        return self._data['Remote']
+
+    @property
+    def RemoteDescriptor(self) -> Optional[Dict]:
+        return self._data.get("RemoteDescriptor", None)
 
     @property
     def ValidityCheck(self) -> Tuple[bool, List[str]]: