+++ Executor/Tasks/Run/__init__.py
@@ -8,3 +8,5 @@ from .publish import Publish
 from .compress_files import CompressFiles
 from .delay import Delay
 from .csvToInflux import CsvToInflux
+from .add_milestone import AddMilestone
+from .publish_from_source import PublishFromPreviousTaskLog, PublishFromFile