+++ REST/rest_client.py
@@ -16,8 +16,8 @@ class Payload(Enum):
 
 class RestClient:
     HEADERS = {'Accept-Language': 'en-US;q=0.5,en;q=0.3'}
-    RETRIES = 10
-    FILENAME_PATTERN = re.compile(r".*filename=\"(.*)\"")
+    RETRIES = 3
+    FILENAME_PATTERN = re.compile(r'.*filename="?(.*)"?')
 
     def __init__(self, api_host, api_port, suffix, https=False, insecure=False):
         self.api_url = f'http{"s" if https else ""}://{api_host}:{api_port}{suffix}'
@@ -34,7 +34,7 @@ class RestClient:
 
     def Trace(self, traceId, url, method, headers=None, body=None, files=None):
         from Helper import Log
-        Log.D(f"[{traceId}] >> [{method}] {self.api_url}/{url}")
+        Log.D(f"[{traceId}] >> [{method}] {url}")
         for name, param in [('Headers', headers), ('Body', body), ('Files', files)]:
             if param is not None:
                 Log.D(f'[{traceId}] >> {name}: {param}')
@@ -52,10 +52,13 @@ class RestClient:
         Log.D(f'[{traceId}] << [Code {code}] {body}')
         return response
 
-    def DownloadFile(self, url, output_folder):
-        response = self.HttpGet(url)
-        filename = self.GetFilename(response.headers["Content-Disposition"])
-        output_file = realpath(join(output_folder, filename))
+    def DownloadFile(self, url, output_folder) -> Optional[str]:
+        try:
+            response = self.HttpGet(url)
+            filename = self.GetFilename(response.headers["Content-Disposition"])
+            output_file = realpath(join(output_folder, filename))
+        except Exception:
+            return None
 
         with open(output_file, 'wb+') as out:
             out.write(response.data)
@@ -69,15 +72,16 @@ class RestClient:
             return result.group(1)
         return "unknown_filename"
 
-    def HttpGet(self, url, extra_headers=None):
+    def HttpGet(self, url, extra_headers=None, timeout=10):
         traceId = self.GetTraceId()
         extra_headers = {} if extra_headers is None else extra_headers
 
         self.Trace(traceId, url, 'GET', headers=extra_headers)
-        return self.DumpResponse(traceId, self.pool.request('GET', url, headers=extra_headers, retries=self.RETRIES))
+        return self.DumpResponse(traceId, self.pool.request('GET', url, headers=extra_headers,
+                                                            retries=self.RETRIES, timeout=timeout))
 
     def HttpPost(self, url, extra_headers=None, body: Optional[Union[str, Dict]] = None,
-                 files=None, payload: Payload = None):
+                 files=None, payload: Payload = None, timeout=10):
         traceId = self.GetTraceId()
         extra_headers = {} if extra_headers is None else extra_headers
 
@@ -101,26 +105,27 @@ class RestClient:
         if files is None:
             return self.DumpResponse(traceId,
                                      self.pool.request('POST', url, body=body or '',
-                                                       headers={**self.HEADERS, **extra_headers}, retries=self.RETRIES))
+                                                       headers={**self.HEADERS, **extra_headers},
+                                                       retries=self.RETRIES, timeout=timeout))
         else:
             return self.DumpResponse(traceId,
                                      post(f"{self.api_url}{url}", data=body, headers={**self.HEADERS, **extra_headers},
                                           files=files, verify=not self.insecure))
 
-    def HttpPatch(self, url, extra_headers=None, body=''):
+    def HttpPatch(self, url, extra_headers=None, body='', timeout=10):
         traceId = self.GetTraceId()
         extra_headers = {} if extra_headers is None else extra_headers
         self.Trace(traceId, url, 'PATCH', headers=extra_headers, body=body)
         return self.DumpResponse(traceId, self.pool.request('PATCH', url, body=body,
                                                             headers={**self.HEADERS, **extra_headers},
-                                                            retries=self.RETRIES))
+                                                            retries=self.RETRIES, timeout=timeout))
 
-    def HttpDelete(self, url, extra_headers=None):
+    def HttpDelete(self, url, extra_headers=None, timeout=10):
         traceId = self.GetTraceId()
         extra_headers = {} if extra_headers is None else extra_headers
         self.Trace(traceId, url, 'DELETE', headers=extra_headers)
         return self.DumpResponse(traceId, self.pool.request('DELETE', url, headers={**self.HEADERS, **extra_headers},
-                                                   retries=self.RETRIES))
+                                                            retries=self.RETRIES, timeout=timeout))
 
     @staticmethod
     def ResponseStatusCode(response) -> int: