+++ Scheduler/dispatcher/routes.py
@@ -1,4 +1,4 @@
-from flask import jsonify, request
+from flask import jsonify, request, abort
 from Status import ExecutionQueue
 from Data import ExperimentDescriptor
 from Scheduler.dispatcher import bp
@@ -7,27 +7,24 @@ from Helper import Log
 
 @bp.route('/run', methods=['POST'])
 def start():
-    data = request.json
-    Log.I("Received execution request")
-    Log.D(f"Payload: {data}")
-    descriptor = ExperimentDescriptor(data)
-    valid, reasons = descriptor.ValidityCheck
+    try:
+        data = request.json
+        Log.I("Received execution request")
+        Log.D(f"Payload: {data}")
 
-    if not valid:
-        executionId = None
-        success = False
-        message = f'Invalid experiment description: {"; ".join(reasons)}'
-    else:
+        if data is None:
+            raise RuntimeError("Received empty payload")
+
+        descriptor = ExperimentDescriptor(data)
+        valid, reasons = descriptor.ValidityCheck
+
+        if not valid:
+            raise RuntimeError(f'Invalid experiment description: {"; ".join(reasons)}')
+
         params = {'Descriptor': descriptor}
         executionId = ExecutionQueue.Create(params).Id
-        success = True
-        message = f'Created execution {executionId}'
-
-    response = {
-        'ExecutionId': executionId,
-        'Success': success,
-        'Message': message
-    }
-
-    Log.D(f"Response: {response}")
-    return jsonify(response)
+        return jsonify({'ExecutionId': executionId})
+    except Exception as e:
+        message = f"Exception while processing execution request: {e}"
+        Log.W(message)
+        return abort(400, message)