+++ Scheduler/east_west/__init__.py
@@ -1,0 +1,5 @@
+from flask import Blueprint
+
+bp = Blueprint('eastWest', __name__)
+
+from Scheduler.east_west import routes