+++ Task/task.py
@@ -14,16 +14,17 @@ class Task:
         self.logMethod = Log.Log if logMethod is None else logMethod
         self.condition = conditionMethod
         self.Vault = {}
+        self.LogMessages = []
 
     def Start(self) -> Dict:
         if self.condition is None or self.condition():
-            self.Log(Level.INFO, f'[Starting Task {self.name}]')
+            self.Log(Level.INFO, f"[Starting Task '{self.name}']")
             self.Log(Level.DEBUG, f'Params: {self.params}')
             self.Run()
-            self.Log(Level.INFO, f'[Task {self.name} finished]')
+            self.Log(Level.INFO, f"[Task '{self.name}' finished]")
             self.Log(Level.DEBUG, f'Params: {self.params}')
         else:
-            self.Log(Level.INFO, f'[Task {self.name} not started (condition false)]')
+            self.Log(Level.INFO, f"[Task '{self.name}' not started (condition false)]")
         return self.params
 
     def Publish(self, key: str, value: object):
@@ -35,3 +36,4 @@ class Task:
 
     def Log(self, level: Union[Level, str], msg: str):
         self.logMethod(level, msg)
+        self.LogMessages.append(msg)